pyproject.toml
[project]
name = "quantumcrew-mail"
version = "1.0.0"
description = "Multi-account Gmail & Calendar MCP server for AI-native inbox management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
 "google-api-python-client>=2.154.0",
 "google-auth>=2.25.0",
 "google-auth-oauthlib>=1.2.0",
 "google-auth-httplib2>=0.2.0",
 "mcp>=1.3.0",
 "pydantic>=2.0.0",
 "python-dotenv>=1.0.1",
 "pytz>=2024.2",
]
[[project.authors]]
name = "QuantumCrew AI"
email = "ai@quantumcrew.ai"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.389",
]

[project.scripts]
quantumcrew-mail = "mcp_gsuite:main"
